% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONSTRU_function.r
\name{constru_Mclust}
\alias{constru_Mclust}
\title{constru_Mclust}
\usage{
constru_Mclust(
  survival_data,
  gene_data,
  prognostic_variable_d,
  cox_formula,
  ncores
)
}
\arguments{
\item{survival_data}{A data frame with the samples as rows and survival time and event as columns.}

\item{gene_data}{A data frame sample names as columns and gene expression as rows.}

\item{prognostic_variable_d}{A vector with the prognostic variable data. It should be of the same length and order as the row of the survival data.}

\item{cox_formula}{Formula used for cox regression. Example: 
\itemize{
  \item Surv( OS_TIME , OS_Event ) ~ prognostic_variable \cr
}
  If additional prognostic factors need to be added to the model, add it as a column in the survival data and use its column name in the formula. Example:
\itemize{
  \item Surv( OS_TIME , OS_Event ) ~ prognostic_variable + AGE \cr
}
  The program will automatically add the interaction term prognostic_variable:gene_data to the formula}

\item{ncores}{The number of cores used during multithreading.}
}
\value{
a table with columns:
\itemize{
  \item mean - gene expression mean
  \item P5\% - 5\% confidence interval of the gene expression
  \item P95\% - 95\% confidence interval of the gene expression
  \item Diff P95\%_5\% - difference between the 95\% and 5\% confidence interval
  \item R - pearson correlation between the gene expression and prognostic variable
  \item nGroups - number of groups found
  \item PS1 - max parity score 1
  \item PS1_max_group - group used to generate PS1
  \item PS1_min_group - group used to generate PS1
  \item PS2 - max parity score 2
  \item PS2_max_group - group used to generate PS2
  \item PS2_min_group - group used to generate PS2
  \item PS3 - max parity score 3
  \item PS3_max_group - group used to generate PS3
  \item PS3_min_group - group used to generate PS3
  \item All_Pval - pvalues of each group
  \item All_HR - Hazard ratio of each group
  \item sample_names - for each sample, a column will be used to indicate which group the sample belongs to
  \item warnings - 
}
}
\description{
Cox regression after separating prognistic variable by gene expression tertiles
}
