% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONSTRU_function.r
\name{constru}
\alias{constru}
\title{constru}
\usage{
constru(clinical, gene_data, metagene_mean, cox_formula, ncores, mode)
}
\arguments{
\item{clinical, }{gene_data, metagene_mean,cox_formula,ncores,mode}
}
\value{
a table detailing ...
}
\description{
Cox regression after separating metagene_mean's tertile category
}
\examples{
todo
}
